Cast floats to int in SetVideoWindow() before comparing them
Same cast is already done when storing those values so in
case there is some rounding done it will cause incorrectly
detected video window size change.

+++ mediaportal/Core/Player/PlaneScene.cs
@@ -371,7 +371,7 @@ namespace MediaPortal.Player
 
         //did the video window,aspect ratio change? if not
         //then we dont need to recalculate and just return the previous settings
-        if (!updateCrop && x == _rectPrevious.X && y == _rectPrevious.Y &&
+        if (!updateCrop && (int)x == _rectPrevious.X && (int)y == _rectPrevious.Y &&
             nw == _rectPrevious.Width && nh == _rectPrevious.Height &&
             GUIGraphicsContext.ARType == _aspectRatioType &&
             GUIGraphicsContext.Overlay == _lastOverlayVisible && _shouldRenderTexture &&
